Cache bounding box using bincode
